+++ app/src/main/java/com/example/playlistmaker/di/AppModule.kt
@@ -2,7 +2,7 @@ package com.example.playlistmaker.di
 
 import com.example.playlistmaker.presentation.playerScreen.PlayerViewModel
 import com.example.playlistmaker.presentation.searchScreen.SearchViewModel
-import com.example.playlistmaker.presentation.mainScreen.MainViewModel
+import com.example.playlistmaker.presentation.rootScreen.RootViewModel
 import com.example.playlistmaker.presentation.settingsScreen.SettingsViewModel
 import org.koin.androidx.viewmodel.dsl.viewModel
 import org.koin.dsl.module
@@ -21,7 +21,7 @@ val appModule = module {
         SettingsViewModel(settingsInteractor = get(), sharingInteractor = get())
     }
 
-    viewModel<MainViewModel> {
-        MainViewModel(settingsInteractor = get())
+    viewModel<RootViewModel> {
+        RootViewModel(settingsInteractor = get())
     }
 }
