--- app/src/main/java/com/example/playlistmaker/presentation/mainScreen/MainActivity.kt
@@ -1,16 +1,0 @@
-package com.example.playlistmaker.presentation.mainScreen
-
-import android.os.Bundle
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.R
-import org.koin.androidx.viewmodel.ext.android.viewModel
-
-class MainActivity : AppCompatActivity() {
-    private val viewModel by viewModel<MainViewModel>()
-    override fun onCreate(savedInstanceState: Bundle?) {
-        super.onCreate(savedInstanceState)
-        setContentView(R.layout.activity_main)
-
-        viewModel.initTheme()
-    }
-}
