+++ app/src/main/java/com/example/playlistmaker/presentation/mediaScreen/MediaFavoritesFragment.kt
@@ -5,19 +5,19 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import com.example.playlistmaker.databinding.FragmentMediaFavoritesContentBinding
+import com.example.playlistmaker.databinding.FragmentMediaFavoritesBinding
 import org.koin.androidx.viewmodel.ext.android.viewModel
 
 class MediaFavoritesFragment : Fragment() {
 
-    private var binding: FragmentMediaFavoritesContentBinding?=null
+    private var binding: FragmentMediaFavoritesBinding? = null
 
     private val playerViewModel by viewModel<MediaFavoritesViewModel>()
     override fun onCreateView(
         inflater: LayoutInflater, container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        binding = FragmentMediaFavoritesContentBinding.inflate(inflater, container, false)
+        binding = FragmentMediaFavoritesBinding.inflate(inflater, container, false)
         return binding?.root
     }
