--- app/src/main/java/com/example/playlistmaker/presentation/searchScreen/SearchActivity.kt
@@ -1,15 +1,0 @@
-package com.example.playlistmaker.presentation.searchScreen
-
-import androidx.appcompat.app.AppCompatActivity
-import com.example.playlistmaker.domain.models.Track
-import com.example.playlistmaker.presentation.ui.ItemClickListener
-
-class SearchActivity : AppCompatActivity(), ItemClickListener, SearchView {
-    override fun render(state: SearchState) {
-        TODO("Not yet implemented")
-    }
-
-    override fun onClick(track: Track) {
-        TODO("Not yet implemented")
-    }
-}
