--- app/src/main/java/com/example/playlistmaker/presentation/settingsScreen/SettingsActivity.kt
@@ -1,6 +1,0 @@
-package com.example.playlistmaker.presentation.settingsScreen
-
-import androidx.appcompat.app.AppCompatActivity
-
-class SettingsActivity : AppCompatActivity() {
-}
