+++ app/src/main/res/layout/activity_root.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".presentation.ui.root.RootActivity">
+    tools:context=".presentation.rootScreen.RootActivity">
 
     <androidx.fragment.app.FragmentContainerView
         android:id="@+id/container_view"
