+++ app/src/main/res/layout/fragment_media_favorites.xml
